# Reimplement semaphore using packaged integer.