**Note:** This needs **thorough** review. It's the 1000th iteration of this and I'm still not 100% certain it's correct :joy: 

<!--
Request Prow to automatically lint any go code in this PR:

/lint
-->

<!-- Please include the 'why' behind your changes if no issue exists -->
## Proposed Changes

This is an implementation of a semaphore based on packed integers and a channel.

state is an int64 that has two int32s packed into it: capacity and inFlight. The former specifies how many request are allowed at any given time into the semaphore while the latter refers to the currently in-flight requests. Packing them both into one int64 allows us to optimize access semantics using atomic operations, which can't be guaranteed on 2 individual values.

The channel is merely used as a vehicle to be able to "wake up" individual goroutines if capacity becomes free. It's not consistently used in accordance to actual capacity but is rather a communication vehicle to ensure waiting routines are properly worken up.

### Benchmarks

```
benchmark                                 old ns/op     new ns/op     delta
BenchmarkBreakerReserve/sequential-16     60.2          23.3          -61.30%
BenchmarkBreakerReserve/parallel-16       243           80.6          -66.83%

benchmark                                    old ns/op     new ns/op     delta
BenchmarkBreakerMaybe/1-sequential-16        92.1          27.5          -70.14%
BenchmarkBreakerMaybe/1-parallel-16          310           224           -27.74%
BenchmarkBreakerMaybe/10-sequential-16       92.1          28.0          -69.60%
BenchmarkBreakerMaybe/10-parallel-16         843           83.1          -90.14%
BenchmarkBreakerMaybe/100-sequential-16      91.6          28.0          -69.43%
BenchmarkBreakerMaybe/100-parallel-16        183           74.0          -59.56%
BenchmarkBreakerMaybe/1000-sequential-16     90.3          27.9          -69.10%
BenchmarkBreakerMaybe/1000-parallel-16       220           74.1          -66.32%
```

**Release Note**

<!-- Enter your extended release note in the below block. If the PR requires
additional action from users switching to the new release, include the string
"action required". If no release note is required, write "NONE". -->

```release-note
NONE
```

/assign @vagababov @julz