@slinkydeveloper 2020-08-12
@markusthoemmes I'm curious, can you share the numbers with memory usage too? `-mem` flag