@markusthoemmes 2020-08-12
@slinkydeveloper I redacted them cause they are the same. 0 allocations throughout.