@slinkydeveloper 2020-08-12
/lgtm <- for the perfomance tweaks, i didn't fully analyzed the concurrency :smile: